+++ tests/test_utils/test_cache.py
@@ -9,6 +9,7 @@ from starlette.responses import PlainTextResponse
 from starlette.types import Scope
 
 from asgi_caches.exceptions import RequestNotCachable, ResponseNotCachable
+from asgi_caches.rules import Rule
 from asgi_caches.utils.cache import (
     deserialize_response,
     get_cache_key,
@@ -41,7 +42,7 @@ async def test_get_from_emtpy_cache(cache: BaseCache, method: str) -> None:
         "headers": [],
     }
     request = Request(scope)
-    response = await get_from_cache(request, cache=cache)
+    response = await get_from_cache(request, cache=cache, rules=[Rule()])
     assert response is None
 
 
@@ -57,7 +58,7 @@ async def test_non_cachable_methods(cache: BaseCache, method: str) -> None:
     }
     request = Request(scope)
     with pytest.raises(RequestNotCachable):
-        await get_from_cache(request, cache=cache)
+        await get_from_cache(request, cache=cache, rules=[Rule()])
 
 
 async def test_store_in_cache(cache: BaseCache) -> None:
@@ -71,7 +72,7 @@ async def test_store_in_cache(cache: BaseCache) -> None:
     response = PlainTextResponse("Hello, world!")
     assert await get_cache_key(request, method="GET", cache=cache) is None
 
-    await store_in_cache(response, request=request, cache=cache)
+    await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
     key = await get_cache_key(request, method="GET", cache=cache)
     assert key is not None
@@ -81,7 +82,7 @@ async def test_store_in_cache(cache: BaseCache) -> None:
 
 
 @pytest.mark.parametrize(
-    "status_code", (201, 202, 204, 301, 307, 308, 400, 401, 403, 500, 502, 503)
+    "status_code", (201, 202, 307, 308, 400, 401, 403, 500, 502, 503)
 )
 async def test_non_cachable_status_codes(cache: BaseCache, status_code: int) -> None:
     scope: Scope = {
@@ -93,7 +94,7 @@ async def test_non_cachable_status_codes(cache: BaseCache, status_code: int) -> 
     request = Request(scope)
     response = PlainTextResponse("Hello, world!", status_code=status_code)
     with pytest.raises(ResponseNotCachable):
-        await store_in_cache(response, request=request, cache=cache)
+        await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
 
 async def test_non_cachable_zero_ttl(cache: BaseCache) -> None:
@@ -110,7 +111,7 @@ async def test_non_cachable_zero_ttl(cache: BaseCache) -> None:
     request = Request(scope)
     response = PlainTextResponse("Hello, world!")
     with pytest.raises(ResponseNotCachable):
-        await store_in_cache(response, request=request, cache=cache)
+        await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
 
 async def test_get_from_cache(cache: BaseCache) -> None:
@@ -122,9 +123,9 @@ async def test_get_from_cache(cache: BaseCache) -> None:
     }
     request = Request(scope)
     response = PlainTextResponse("Hello, world!")
-    await store_in_cache(response, request=request, cache=cache)
+    await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
-    cached_response = await get_from_cache(request, cache=cache)
+    cached_response = await get_from_cache(request, cache=cache, rules=[Rule()])
     assert cached_response is not None
     assert ComparableStarletteResponse(cached_response) == response
     assert "Expires" in cached_response.headers
@@ -140,9 +141,9 @@ async def test_default_max_age(cache: BaseCache) -> None:
     }
     request = Request(scope)
     response = PlainTextResponse("Hello, world!")
-    await store_in_cache(response, request=request, cache=cache)
+    await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
-    cached_response = await get_from_cache(request, cache=cache)
+    cached_response = await get_from_cache(request, cache=cache, rules=[Rule()])
     assert cached_response is not None
     now = dt.datetime.now(tz=dt.timezone.utc)
     http_date_fmt = "%a, %d %b %Y %H:%M:%S GMT"
@@ -164,9 +165,9 @@ async def test_cache_ttl_max_age(short_cache: BaseCache) -> None:
     }
     request = Request(scope)
     response = PlainTextResponse("Hello, world!")
-    await store_in_cache(response, request=request, cache=short_cache)
+    await store_in_cache(response, request=request, cache=short_cache, rules=[Rule()])
 
-    cached_response = await get_from_cache(request, cache=short_cache)
+    cached_response = await get_from_cache(request, cache=short_cache, rules=[Rule()])
     assert cached_response is not None
     now = dt.datetime.now(tz=dt.timezone.utc)
     http_date_fmt = "%a, %d %b %Y %H:%M:%S GMT"
@@ -188,9 +189,9 @@ async def test_get_from_cache_head(cache: BaseCache) -> None:
     }
     request = Request(scope)
     response = PlainTextResponse("Hello, world!")
-    await store_in_cache(response, request=request, cache=cache)
+    await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
-    cached_response = await get_from_cache(request, cache=cache)
+    cached_response = await get_from_cache(request, cache=cache, rules=[Rule()])
     assert cached_response is not None
     assert ComparableStarletteResponse(cached_response) == response
 
@@ -204,11 +205,11 @@ async def test_get_from_cache_different_path(cache: BaseCache) -> None:
     }
     request = Request(scope)
     response = PlainTextResponse("Hello, world!")
-    await store_in_cache(response, request=request, cache=cache)
+    await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
     other_scope = {**scope, "path": "/other_path"}
     other_request = Request(other_scope)
-    cached_response = await get_from_cache(other_request, cache=cache)
+    cached_response = await get_from_cache(other_request, cache=cache, rules=[Rule()])
     assert cached_response is None
 
 
@@ -223,11 +224,11 @@ async def test_get_from_cache_vary(cache: BaseCache) -> None:
     # Response indicates that contents of the response at this URL may *vary*
     # depending on the "Accept-Encoding" header sent in the request.
     response = PlainTextResponse("Hello, world!", headers={"Vary": "Accept-Encoding"})
-    await store_in_cache(response, request=request, cache=cache)
+    await store_in_cache(response, request=request, cache=cache, rules=[Rule()])
 
     # Let's use a different "Accept-Encoding" header,
     # and check that no cached response is found.
     other_scope = {**scope, "headers": [[b"accept-encoding", b"identity"]]}
     other_request = Request(other_scope)
-    cached_response = await get_from_cache(other_request, cache=cache)
+    cached_response = await get_from_cache(other_request, cache=cache, rules=[Rule()])
     assert cached_response is None